partially added night time
started adding night time. at the moment if you open the inventory, the
night time tiles are rendered. yes there is a small issue with it since
the same tiles in the dungeon are the same as outdoor, so they get
darker too, that will be fixed. The next step is to pull time from some
where so it happens on it's own.