NPC interaction started
there are two npc's on the northern most part of the beach.
use spacebar to interact with them. they are each their own unique npc.

each npc file has it's own List<String> for the text that it displays.

it also has it's own String for the type of NPC that it is

I'm abandoning scaling the window, it's too much work for too little
reward. it will just end up being one size. the height is equivalent to
the top frame, + 40x the middle piece, + the bottom frame. I didn't have
enough time to do the math.