Quest
the quests are now stored properly in the engine upon startup.

next step is to make them save periodically and once more on exit, this
should help prevent as many issues as possible with quests due to loss
of data on crashes, etc. Granted backups will still need to be made
regularly once the game is operational, no way around that ... shit just
happens.